virt: Move PciAssignable out of virt_utils module
This PciAssignable class depends kvm_control module which will raise
an exception in guest. Thus move this class to virt_test_setup module
so that we can use virt_utils in client tests.

Acked-by: Yiqiao Pu <ypu@redhat.com>
